refactor: xmodule/annotatable_module.py -> xmodule/annotatable_block.py
