+++ xmodule/static_content.py
@@ -19,7 +19,7 @@ import django
 from docopt import docopt
 from path import Path as path
 
-from xmodule.annotatable_module import AnnotatableBlock
+from xmodule.annotatable_block import AnnotatableBlock
 from xmodule.capa_module import ProblemBlock
 from xmodule.conditional_module import ConditionalBlock
 from xmodule.html_module import AboutBlock, CourseInfoBlock, HtmlBlock, StaticTabBlock
